% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unisens.R
\name{read_props}
\alias{read_props}
\title{Read unisens sample properties}
\usage{
read_props(folder)
}
\arguments{
\item{folder}{path to the unisens data folder}
}
\value{
list of attributes associated with the sample
}
\description{
Parses the xml file into a list to extract the attributes (such as starting
timestamp) of the sample.
}
