% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unisens.R
\name{read_unisens}
\alias{read_unisens}
\title{Read unisens binary files}
\usage{
read_unisens(folder, file_id)
}
\arguments{
\item{folder}{the folder containing the xml file and the .bin file(s)}

\item{file_id}{the filename of the binary file. This should also exist in the
xml file as an ID. Upon error, a nice warning will be given.}
}
\value{
A unisens data matrix. The named columns of the matrix are the
  different channels, and the rows of the matrix are the samples. The
  attributes of the matrix will include further information about the file,
  such as the sample rate.
}
\description{
Function to read unisens binary files into a matrix. Reading happens in
blocks of 1 000 000 values at a time, to keep track of progress and to keep
the process manageable in size. This function will display a nice
progress bar in interactive use.
}
\details{
The reading will happen in blocks of 1 000 000 values at a time, to
  keep track of progress and to keep the process manageable in size.
}
\seealso{
\code{\link{sample_rate}}, \code{\link{sub_sample}}
}
