% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unisens.R
\name{sample_rate}
\alias{sample_rate}
\title{Check the sample rate of unisens data.}
\usage{
sample_rate(uni_mat)
}
\arguments{
\item{uni_mat}{a data matrix as returned by the read_unisens function}
}
\description{
Check the sample rate of unisens data.
}
