% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unisens.R
\name{sub_sample}
\alias{sub_sample}
\title{Subsample unisens data to a target sample rate}
\usage{
sub_sample(uni_mat, tgt_rate)
}
\arguments{
\item{uni_mat}{a data matrix as returned by the read_unisens function}

\item{tgt_rate}{a numeric value containing the target sample rate in Hz}
}
\value{
a unisens data matrix
}
\description{
This function subsamples a unisens data matrix to match a target sample rate.
The target sample rate in Hz is approximately matched by selecting the rows
closest to the target sample rows.
}
